requirements.txt
Django>=3.1,<3.2
djangorestframework>=3.12.4,<3.13.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
psycopg2>=2.8.6,<2.9.0
Pillow>=6.2.1,<6.3.0

flake8>=3.9.1,<4.0.0